chore: int64,uint64 -> int,uint

+++ plugins/v4/jobs/job.go
@@ -13,7 +13,7 @@ type Queue interface {
 	// ExtractMin returns the item with the highest priority (less value is the highest priority)
 	ExtractMin() Job
 	// Len returns the number of items in the queue
-	Len() uint64
+	Len() uint
 }
 
 // Job represents a binary heap item
@@ -44,12 +44,12 @@ type Message interface {
 	// Payload returns the data associated with the job
 	Payload() []byte
 	// Delay returns the delay time for the Job (not supported by all drivers)
-	Delay() int64
+	Delay() int
 	// AutoAck returns the autocommit status for the Job
 	AutoAck() bool
 	// UpdatePriority sets the priority of the Job. Priority is optional but cannot be set to 0.
 	// The default priority is 10
-	UpdatePriority(int64)
+	UpdatePriority(int)
 	// Headers returns the metadata for the item
 	Headers() map[string][]string
 }
@@ -57,9 +57,9 @@ type Message interface {
 // KAFKA options (leave them empty for other drivers)
 type KafkaOptions interface {
 	// Offset returns the offset associated with the Job
-	Offset() int64
+	Offset() int
 	// Partition returns the partition associated with the Job
-	Partition() int32
+	Partition() int
 	// Topic returns the topic associated with the Job
 	Topic() string
 	// Metadata returns the metadata associated with the Job
@@ -85,7 +85,7 @@ type Pipeline interface {
 	// This might be used for SQS attributes or tags, for example
 	Map(name string, out map[string]string) error
 	// Priority returns the default pipeline priority
-	Priority() int64
+	Priority() int
 	// Get is used to retrieve the data associated with a key
 	Get(key string) interface{}
 }
